refactor: move prisma init to src/db/
